# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Bolo-Man API
  description: Cameroon-focused Daily Services Platform API
  version: '1.0'
  contact:
    name: Bolo-Man Support
    email: support@boloman.cm
servers:
  - url: https://api.boloman.cm
    description: Production
  - url: http://localhost:3000
    description: Development
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      tags: [services]
      summary: List services
      parameters:
        - name: categoryId
          in: query
          schema: { type: string }
      responses:
        200:
          description: List of services
  /providers/{providerId}/services:
    get:
      tags: [services]
      summary: Get services by provider
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Provider services
...
  /subscriptions:
    get:
      tags: [subscriptions]
      summary: List subscription plans
      security: []
      parameters:
        - name: countryCode
          in: query
          schema: { type: string, default: 'CM' }
      responses:
        200:
          description: List of plans

    post:
      tags: [subscriptions]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscriptionId: { type: string }
                paymentMethod: { type: string }

    put:
      tags: [subscriptions]
      summary: Upgrade/downgrade subscription
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                subscriptionId: { type: string }

    delete:
      tags: [subscriptions]
      summary: Cancel subscription
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...